Modified instance to use doubles instead of float. This will increase the memory usage, but will also increase ease of use. This update will most likely break all backward compatibility, but the library is not yet stable and we feel that this change has more benefits than drawbacks.
git-svn-id: https://svn.code.sf.net/p/java-ml/code/trunk@679 3162d2cb-3e1f-0410-96b0-914c58786302

+++ junit/clustering/IterativeKMeansTest.java
@@ -78,7 +78,7 @@ public class IterativeKMeansTest {
     }
     
     private void add(Dataset data,float x,float y){
-        float[] values={x,y};
+        double[] values={x,y};
         SimpleInstance in=new SimpleInstance(values);
         data.addInstance(in);
     }
