+++ junit/clustering/KMedoidsTest.java
@@ -86,7 +86,7 @@ public class KMedoidsTest {
     }
     
     private void add(Dataset data,float x,float y){
-        float[] values={x,y};
+        double[] values={x,y};
         SimpleInstance in=new SimpleInstance(values);
         data.addInstance(in);
     }
