+++ junit/clustering/SOMTest.java
@@ -80,7 +80,7 @@ public class SOMTest {
     }
     
     private void add(Dataset data,float x,float y){
-        float[] values={x,y};
+        double[] values={x,y};
         SimpleInstance in=new SimpleInstance(values);
         data.addInstance(in);
     }
