+++ net/sf/javaml/filter/NormalizeMean.java
@@ -76,7 +76,7 @@ public class NormalizeMean implements Filter {
 	public Instance filterInstance(Instance instance) {
 		if(mean==null||std==null)
 			throw new RuntimeException("You should first call filterDataset for this filter, some parameters are not yet set.");
-		float[] out=new float[instance.size()];
+        double[] out=new double [instance.size()];
 		for(int i=0;i<out.length;i++){
 			out[i]=(instance.getValue(i)-mean[i])/std[i];
 		}
@@ -91,7 +91,7 @@ public class NormalizeMean implements Filter {
     public Instance unfilterInstance(Instance instance) {
         if(mean==null||std==null)
             throw new RuntimeException("You should first call filterDataset for this filter, some parameters are not yet set.");
-        float[] out=new float[instance.size()];
+        double[] out=new double[instance.size()];
         for(int i=0;i<out.length;i++){
             //out[i]=(instance.getValue(i)-mean[i])/std[i];
             out[i]=(instance.getValue(i)*std[i])+mean[i];
